refactor search-zone/search-zone ->search-zone/index
