+++ frontend/src/components/(game)/search-zone/index.tsx
