+++ frontend/src/contexts/(game)/search/drawer.tsx
@@ -6,7 +6,7 @@
 // } from "@/components/(ui)/drawer";
 // import { VisuallyHidden } from "@radix-ui/react-visually-hidden";
 import { useSearchContext } from ".";
-import SearchZone from "@/components/(game)/search-zone/search-zone";
+import SearchZone from "@/components/(game)/search-zone";
 import { useEffect, useState } from "react";
 import { ShahrazadZoneId } from "@/types/bindings/zone";
 // import { Button } from "@/components/(ui)/button";
